+++ src/query/service/src/interpreters/interpreter_database_show_create.rs
@@ -18,7 +18,7 @@ use std::sync::Arc;
 use common_datablocks::DataBlock;
 use common_datavalues::prelude::*;
 use common_exception::Result;
-use common_legacy_planners::ShowCreateDatabasePlan;
+use common_planner::plans::ShowCreateDatabasePlan;
 
 use crate::interpreters::Interpreter;
 use crate::pipelines::PipelineBuildResult;
