+++ src/query/service/src/interpreters/interpreter_database_undrop.rs
@@ -15,7 +15,7 @@
 use std::sync::Arc;
 
 use common_exception::Result;
-use common_legacy_planners::UndropDatabasePlan;
+use common_planner::plans::UndropDatabasePlan;
 
 use crate::interpreters::Interpreter;
 use crate::pipelines::PipelineBuildResult;
