+++ src/query/service/src/sql/planner/binder/ddl/database.rs
@@ -30,14 +30,14 @@ use common_datavalues::DataSchemaRefExt;
 use common_datavalues::ToDataType;
 use common_datavalues::Vu8;
 use common_exception::Result;
-use common_legacy_planners::ShowCreateDatabasePlan;
-use common_legacy_planners::UndropDatabasePlan;
 use common_meta_app::schema::DatabaseMeta;
 use common_meta_app::share::ShareNameIdent;
 use common_planner::plans::CreateDatabasePlan;
 use common_planner::plans::DropDatabasePlan;
 use common_planner::plans::RenameDatabaseEntity;
 use common_planner::plans::RenameDatabasePlan;
+use common_planner::plans::ShowCreateDatabasePlan;
+use common_planner::plans::UndropDatabasePlan;
 
 use crate::sql::binder::Binder;
 use crate::sql::planner::semantic::normalize_identifier;
