+++ src/query/service/src/sql/planner/plans/mod.rs
@@ -74,17 +74,17 @@ use common_legacy_planners::RenameTablePlan;
 use common_legacy_planners::RevokePrivilegePlan;
 use common_legacy_planners::RevokeRolePlan;
 use common_legacy_planners::SettingPlan;
-use common_legacy_planners::ShowCreateDatabasePlan;
 use common_legacy_planners::ShowCreateTablePlan;
 use common_legacy_planners::ShowGrantsPlan;
 use common_legacy_planners::TruncateTablePlan;
-use common_legacy_planners::UndropDatabasePlan;
 use common_legacy_planners::UndropTablePlan;
 use common_legacy_planners::UseDatabasePlan;
 use common_planner::plans::CallPlan;
 use common_planner::plans::CreateDatabasePlan;
 use common_planner::plans::DropDatabasePlan;
 use common_planner::plans::RenameDatabasePlan;
+use common_planner::plans::ShowCreateDatabasePlan;
+use common_planner::plans::UndropDatabasePlan;
 use common_planner::MetadataRef;
 pub use copy_v2::CopyPlanV2;
 pub use copy_v2::ValidationMode;
